requirements.txt
streamlit>=1.20.0
mysql-connector-python>=8.0.33
pandas>=1.5.3
numpy>=1.24.2
matplotlib>=3.7.1
reportlab>=3.6.12
streamlit-option-menu>=0.5.2
openpyxl>=3.1.2